connect microservices

+++ order-service/src/main/java/uz/nt/orderservice/controller/OrderController.java
@@ -30,7 +30,7 @@ public class OrderController {
     }
 
     @PostMapping
-    public ResponseDto addOrderIfNotExistUserOrders(@RequestParam Integer product_id, @RequestParam Integer amount){
+    public ResponseDto addOrderIfNotExistUserOrders(@RequestParam Integer product_id, @RequestParam Double amount){
         return orderService.addOrderIfNotExistUserOrders(product_id, amount);
     }
